# src/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Turath GPT Proxy
  version: 1.0.0
paths:
...
  /books/{book_id}/pages/{page}:
    get:
      operationId: getBookPage
      parameters:
        - in: path
          name: book_id
          required: true
          schema: { type: integer }
        - in: path
          name: page
          required: true
          schema: { type: integer }
      responses: { "200": { description: ok } }
...